Crate tf_rosrust

This is a rust port of the ROS tf library. It is intended for being used in robots to help keep track of multiple coordinate frames and is part of a larger suite of rust libraries that provide support for various robotics related functionality.

Example usage:

use tf_rosrust::TfListener;

rosrust::init("listener");
let listener = TfListener::new();

let rate = rosrust::rate(1.0);
while rosrust::is_ok() {
    let tf = listener.lookup_transform("camera", "base_link", rosrust::Time::new());
    println!("{tf:?}");
    rate.sleep();
}

Re-exports

pub use transforms::geometry_msgs::TransformStamped;

Modules

transforms

Structs

TfBroadcaster

Broadcast tf messages

TfBuffer

TfListener

This struct tries to be the same as the C++ version of TransformListener. Use this struct to lookup transforms.

Enums

TfError

Enumerates the different types of errors